Refactor Game mode/state; Music strategies take Composer State;
Two changes here, apologies. The first is a refactor of Game mode/state, where we remove the individual methods to set intervals.

Now that we have intervals set as a index of a chromatic scale, we represent this structure in game state, and functions to update in Game mode.

The other change is to update Musical Strategy interface to take in a ref to the Composer state, making it easier for the strategy to pull the data it needs in its concrete classes.