WIP: Remove FInstrumentScheduleData and FMusicalStrategy structs
Status: Compiles, untested.

In an attempt to reduce the logical complexity of the Composer, this change flattens the FInstrumentScheduleData and FMusicalStrategy objects.

The interaction between ScheduleData and MusicalStrategy were unclear to me. By flattening them, I hope to reduce the overall complexity of the data objects being managed, and have a clearer idea of exactly what data objects are strictly needed.

As part of this, I've also started using Shared Pointers for USTRUCTs more. The thought being that a reference to an object is probably better to use than a copy of the object.

This is compiling, but now needs a bunch of testing.