WIP: Make ComposerData and InstrumentSchedules `TSharedPtr`
My original understanding of TSharedPtr was incorrect. I had thought there should be one true owner of the object, then things sharing the pointer to the object.

The intended usage of TSharedPtr is to have _everything_ be TSharedPtr, no owner of the actual object.

This is still crashing at runtime. I'm chasing down the errors as they appear, as this is a fairly major change. But, with shared pointers I will finally be reading and writing to the correct thing.