Add chance to rest per beat
One thing I want to avoid is having a note be selected on every beat for every bar. The notes you don't play are often as important as the ones you do.

I added a very simple random selection from 0-100, checking if the selection is greater than a derived chance to rest.