Weight candidate notes by musical tension
This is untested, but the idea is almost there.

For each possible note, we need to determine how much musical tension its selection would bring.

The tricky part is that some tension is desired, but not too much. How much tension, and when that tension is desired... all good questions attempted to be answered in this function.