Increment tension weights per beat
Candidate Tension is now the base weight, with bonuses applied to the weight if within tension, etc.

These values are arbitrary, and are essentially magic numbers that could/should be tuned. An argument could be made to pull these tunable values into ComposerState.

But the main work here is to pull Tension/Resolution values into the struct renamed `FNoteCandidate`.

This is needed to increment values outside one of the for loops. This use of struct feels kinda janky but seems to make sense.