Improve chance of choosing different MusicalStrategies
Using the same idea as in CreateMotif, I updated ChooseMusicalStrategy to convert a strategies Appropriateness into a range of ints. If a random range int is within the range defined, it will be the one chosen.

This change means that the highest weighted strategy has a better chance of being chosen, whereas before it was always the one being chosen.