Extract CreateMotif data properties into struct
The debugger was acting weird on a couple properties being set, the cause of which was a switch statement missing breaks.

Anyways, I discovered that once I moved everything into a struct, which is something I wanted to do anyways.

Testing has proceeded to CreateMotif debugging, which is proceeding decently.